Fix that randomly fail to access PSU by btools.py, RESTful, etc.
1. Make sure that to control the PCA9548 should include 'open'
and 'close' channels.
2. Remove 'psumuxmon' daemon because of no ltc4151 of PSUs. With
it, easily mess up the PCA9548 operation.
3. The communication between BMC and PSU would randomly disobey
the SMBus/PMBus protocol - "Devices participating in a transfer
will timeout when any clock low exceeds the value of TTIMEOUT,MIN
of 25 ms.", so it causes the whole PSU-related access operations
go wrong. Add retry mechanism for psu driver to workaround it
temporarily. To upgrade Kernel is required in the future.
4. The 1st 'error status' bit of RESTful api should be a integer
not a char.